fix(energy): 修正能量计算中的类型转换错误
- 在计算总能量时将能量值强制转换为long类型
- 确保乘法操作中使用long类型避免溢出
- 修复了基于itemAccess获取能量总量的逻辑错误

+++ src/main/java/dev/dubhe/anvilcraft/api/energy/ItemFEStorage.java
@@ -97,7 +97,7 @@ public class ItemFEStorage implements EnergyHandler {
     @Override
     public long getAmountAsLong() {
         if (this.itemAccess != null) {
-            return (long) this.itemAccess.getAmount() * this.getEnergyFrom(this.itemAccess.getResource());
+            return (long) this.itemAccess.getAmount() * (long) this.getEnergyFrom(this.itemAccess.getResource());
         }
         if (this.stack == null) return 0;
         return this.stack.getOrDefault(ModComponents.STORED_ENERGY, StoredEnergy.EMPTY).value();
@@ -108,7 +108,7 @@ public class ItemFEStorage implements EnergyHandler {
         if (this.itemAccess != null) {
             ItemResource resource = this.itemAccess.getResource();
             if (!resource.is(this.validItem)) return 0;
-            return (long) this.itemAccess.getAmount() * this.capacity;
+            return (long) this.itemAccess.getAmount() * (long) this.capacity;
         }
         return this.capacity;
     }
